requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-pinecone==0.0.3
pinecone-client==3.0.0
openai==1.6.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
jinja2==3.1.2
orjson==3.9.10
httpx==0.25.2

# Dev / Testing tools
pytest==7.4.3
black==23.12.1
ruff==0.1.7
pre-commit==3.5.0
